Remove selectedData method Put scoreboard data selector in separate import
